refactor: improve code with new Search component that handles both types of search
