--- Screens/SearchByCity.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { View, Text } from 'react-native'
-
-export default function SearchByCity() {
-  return (
-    <View>
-      <Text>Search by city</Text>
-    </View>
-  )
-}
